requirement.txt
pandas
pyodbc
openpyxl
python-dotenv